### 2022-08-30 #4700 @leouieda
👋🏾  Hi @dshean, I'll be handling the review process for your paper.

This is a pre-review where we make sure everything is alright and try to find reviewers. The actual review will only start later on and will take place in a different issue. I'll do a first editorial pass through the submission to make sure it fits our submission requirements (I may ask for a few clarifications or modifications). If all is well, I'll proceed to inviting reviewers and once we have enough (usually 2-4) we'll start the actual review.

If you have any questions or concerns, please feel free to ping me in this issue or email me if the message needs to be private.

Please keep an eye on your GitHub notifications since this will be our primary mode of communication.
