### 2022-08-30 #4700 @leouieda
@dshean is this submission about just the Python client (`sliderule-python`) or about the entire stack (server, client, and icesat-2 plugin)? The paper is describing the entire stack but the repository linked above that hosts the paper is just the Python client.
