### 2022-08-31 #4700 @leouieda
@dshean it's not an issue 🙂 I just wanted to get a better picture to figure out how to handle the reviewer invitations.

The JOSS review is focused on the software (code, documentation, community practices) and so it needs to be clear which code repositories are being reviewed. Since this is a very integrated stack with server and client apps, I think it makes sense to review it all at once. The client by itself might not be enough to warrant a separate publication. I'll try to get reviewers with skills that can cover each aspect of the stack and the domain application for ICESat-2.

Since the code spans multiple repositories, you'll need to make a few modification to the submission before we can get started:

1. Create a new separate repository in your organisation to host the JOSS `paper.md`.
3. Post a link to the new paper repo here.
4. Include links in the paper text and the new repo's README to the individual repositories for the server, client, and any other code repository included in the review. This is to make it clear to reviewers and readers of the paper which code was reviewed.

This is how we've handled multiple repository submissions in the past and it works well to lower the complexity for reviewers.

Once this is done, I will start inviting reviewer. If you have any suggestions for possible reviewers, post their user names or email. **Please don't @-mention them** (e.g. @leouieda) to avoid notifying them ahead of an invitation from me.
