+++ README.md
@@ -1,0 +1,2 @@
+# paper
+JOSS paper materials
