### 2022-08-31 0bff8d9
Moved submission materials from sliderule-python repo
