+++ paper/paper.bib
@@ -1,0 +1,178 @@
+@article{neumann2019ice,
+  title={The Ice, Cloud, and Land Elevation Satellite--2 Mission: A global geolocated photon product derived from the advanced topographic laser altimeter system},
+  author = {Neumann, T. A. and Anthony J. Martino and Thorsten Markus and Sungkoo Bae and Megan R. Bock and Anita C. Brenner and Kelly M. Brunt and John Cavanaugh and Stanley T. Fernandes and David W. Hancock and Kaitlin Harbeck and Jeffrey Lee and Nathan T. Kurtz and Philip J. Luers and Scott B. Luthcke and Lori Magruder and Teresa A. Pennington and Luis Ramos-Izquierdo and Timothy Rebold and Jonah Skoog and Taylor C. Thomas},
+  journal={Remote Sensing of Environment},
+  volume={233},
+  pages={111325},
+  year={2019},
+  publisher={Elsevier},
+  doi={10.1016/j.rse.2019.111325}
+}
+
+@article{neumann2019atbd,
+  title={Ice, cloud, and land elevation satellite--2 (ICESat-2) project: Algorithm theoretical basis document (ATBD) for global geolocated photons (ATL03)},
+  author = {Neumann, T. A. and A. Brenner and D. Hancock and J. Robbins and J. Saba and K. Harbeck and A. Gibbons and J. Lee and S. B. Luthcke and T. Rebold},
+  journal={National Aeronautics and Space Administration, Goddard Space Flight Center},
+  year={2019},
+  url={https://nsidc.org/sites/nsidc.org/files/technical-references/ICESat2_ATL03_ATBD_r005.pdf}
+}
+
+@article{markus2017ice,
+title = {The Ice, Cloud, and land Elevation Satellite-2 (ICESat-2): Science requirements, concept, and implementation},
+journal = {Remote Sensing of Environment},
+volume = {190},
+pages = {260-273},
+year = {2017},
+issn = {0034-4257},
+doi = {10.1016/j.rse.2016.12.029},
+url = {https://www.sciencedirect.com/science/article/pii/S0034425716305089},
+author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalati and Kelly Brunt and Beata Csatho and Sinead Farrell and Helen Fricker and Alex Gardner and David Harding and Michael Jasinski and Ron Kwok and Lori Magruder and Dan Lubin and Scott Luthcke and James Morison and Ross Nelson and Amy Neuenschwander and Stephen Palm and Sorin Popescu and CK Shum and Bob Schutz and Benjamin Smith and Yuekui Yang and Jay Zwally}
+}
+
+@article{smith2019land,
+  title={Land ice height-retrieval algorithm for NASA's ICESat-2 photon-counting laser altimeter},
+  author={Smith, Benjamin and Fricker, Helen A and Holschuh, Nicholas and Gardner, Alex S and Adusumilli, Susheel and Brunt, Kelly M and Csatho, Beata and Harbeck, Kaitlin and Huth, Alex and Neumann, Thomas and others},
+  journal={Remote Sensing of Environment},
+  volume={233},
+  pages={111352},
+  year={2019},
+  publisher={Elsevier},
+  doi={10.1016/j.rse.2019.111352}
+}
+
+@article{neuenschwander2019atl08,
+  title={The ATL08 land and vegetation product for the ICESat-2 Mission},
+  author={Neuenschwander, Amy and Pitts, Katherine},
+  journal={Remote sensing of environment},
+  volume={221},
+  pages={247--259},
+  year={2019},
+  publisher={Elsevier},
[…]
