+++ paper/paper.md
@@ -1,0 +1,107 @@
+---
+title: 'SlideRule: Enabling rapid, scalable, open science for the NASA ICESat-2 mission and beyond'
+tags:
+  - Python
+  - NASA
+  - ICESat-2
+  - Glacier
+  - Earth observation
+authors:
+  - name: David Shean ^[co-first author] 
+    orcid: 0000-0003-3840-3860
+    affiliation: 1
+  - name: J.P. Swinski ^[co-first author] # note this makes a footnote saying 'co-first author'
+    affiliation: 2
+  - name: Ben Smith
+    affiliation: 3
+  - name: Tyler Sutterley
+    affiliation: 3
+  - name: Scott Henderson
+    orcid: 0000-0003-0624-4965
+    affiliation: "4, 5"
+  - name: Carlos Ugarte
+    affiliation: 2
+  - name: Thomas Neumann
+    affiliation: 2
+affiliations:
+ - name: University of Washington, Department of Civil and Environmental Engineering
+   index: 1
+ - name: NASA Goddard Space Flight Center
+   index: 2
+ - name: University of Washington, Applied Physics Laboratory
+   index: 3
+ - name: University of Washington, Department of Earth and Space Sciences
+   index: 4
+ - name: University of Washington, eScience Institute
+   index: 5
+
+date: 19 August 2022
+bibliography: paper.bib
+---
+
+# Summary
+`SlideRule` is an open source server-side framework for on-demand processing of science data in the cloud. The `SlideRule` project offers a new paradigm for NASA archival data management – rapid delivery of customizable on-demand data products, rather than hosting large volumes of standard derivative products, which will inevitably be insufficient for some science applications. 
+
+The scalable server-side components of `SlideRule` run in the AWS cloud with optimized functions to read HDF5 data hosted by NASA in S3 cloud object storage. While `SlideRule` can be accessed by any HTTP client (e.g., curl) through GET and POST requests, the `sliderule-python` client provides a user-friendly API for easy interaction with the `SlideRule` service. The client library returns standard Python data containers (i.e., Pandas DataFrame) and facilitates serialization with provenance metadata for reproducible science.
+
+`SlideRule` uses a plugin framework to support different NASA missions and data products. The ICESat-2 `SlideRule` plugin offers customizable algorithms to process the archive of low-level data products from the NASA Ice Cloud and land Elevation Satellite-2 (ICESat-2) laser altimetry mission. The user defines a geographic area of interest and key processing parameters via an interactive web interface or the API, and SlideRule returns high-level surface elevation point cloud products in seconds to minutes, enabling rapid algorithm developent, visualization and scientific interpretation. 
+
+# Statement of need
+ICESat-2 launched in September 2018 as the follow-on mission for the original ICESat mission (2003-2010), which pioneered the application of space-based laser altimetry to precisely measure the height of the Earth's changing surface [@markus2017ice]. These missions are optimized to measure elevation change of the cryosphere, including ice sheets and sea ice, allowing scientists to quantify and understand changes in a warming climate.
+
+ICESat-2 carries the Advanced Topographic Laser Altimeter System (ATLAS), which records the two-way travel time of transmitted photons from six beams [@neumann2019ice]. The Level 2A Geolocated Photon Data Product (ATL03) provides unique latitude, longitude, height and timing of each photon, along with many other attributes, including basic classification as a return from the Earth's surface vs. background photons from sunlight or instrument noise.
+
[…]
