### 2022-09-02 0bdca54
Update paper with project repo organization section
