### 2022-09-02 b70b2e8
Minor paper updates
