### 2022-09-02 #4700 @dshean
Link to new paper repo: https://github.com/ICESat2-SlideRule/paper

README.md and paper.md both contain updated list of project repos with links and brief description
