### 2022-09-02 #4700 @dshean
@leouieda, here is a list of potential reviewers, with links to github pages. We attempted to identify individuals with broad expertise who could review different components of the SlideRule project:
- [Joe Kennedy](https://github.com/jhkennedy), ASF/HyP3
- [Alex Gardner](https://github.com/alex-s-gardner), JPL/ITS_LIVE
- [Chris Crosby](https://github.com/cjcrosby), OpenTopography/OpenAltimetry
- [Howard Butler](https://github.com/hobu), Hobu/PDAL
- [Jessica Scheick](https://github.com/JessicaS11), icepyx
- [John Readey](https://github.com/jreadey), HDFGroup
- [Luis Lopez](https://github.com/betolink), NSIDC
