### 2022-10-06 #4700 @dshean
Hi @leouieda !  Any update on review status?  We can recommend additional reviewers if necessary.

We are getting requests from users on they should cite SlideRule for their publications using the service and resulting data products.  For now we can point to Zenodo records, but hoping to have the JOSS paper out as the primary source.  Thanks!
