### 2022-10-16 #4700 @arfon
:wave: @cjcrosby @hobu @JessicaS11 – would any of you be willing to review this submission for JOSS? The submission under consideration is SlideRule, a toolset for researchers leveraging the NASA ICESat-2 mission.

Many thanks
Arfon
