### 2022-10-20 #4700 @cjcrosby
Thanks for the invitation, but I have some upcoming PTO and other pressing projects that will prevent me from providing a timely review.

Off the top of my head, these are some other folks who may be interested in reviewing:
Laurence Hawker, University of Bristol (https://scholar.google.com/citations?user=k5ajFzUAAAAJ&hl=en) who used ICESat-2 for the new FABDEM product
Tom Hengl, OpenGeoHub (https://opengeohub.org/poeple/tom-hengl/)
John Lindsey, University of Guelph / Whitebox Geospatial (https://jblindsay.github.io/ghrg/software.shtml)
