### 2022-10-24 #4700 @JessicaS11
I have an affiliate appointment at the University of Washington, the lead submitter's institution. He and I worked together organizing Hackweeks (and are thus also on some joint JupyterBook Zenodo releases). I have also attended many SlideRule developer meetings, am on their Slack page, and occasionally meet with their lead developer to work on synergies between our two tools ([icepyx])(https://github.com/icesat2py/icepyx).
