### 2022-11-07 #4700 @dshean
Hi @danielskatz, any update on reviewers for this paper?  This was submitted August 22, 2022.  We can recommend additional reviewers if needed.

I note that @JessicaS11 now has a JOSS paper under review for an October 3, 2022 submission, describing another ICESat-2 software package: https://github.com/openjournals/joss-reviews/issues/4912

If the scope of our submission is too large, we can limit review to the `sliderule-python` repository.  Thanks!
