### 2022-11-08 #4700 @danielskatz
👋 @dshean - we've gone to a track-based model of editing, and the editor for this track is @openjournals/ese-eics who I am pinging here
