### 2022-11-08 #4700 @kthyng
Hi @dshean and thanks for your patience. I'll be taking over as editor. I'll do my best to get this rolling, I'm sorry for the delay!
