### 2022-11-08 #4700 @kthyng
Hi @jhkennedy and @jreadey! Would any of you be willing to review this submission for JOSS? The submission under consideration is SlideRule, a toolset for researchers leveraging the NASA ICESat-2 mission.
