### 2022-11-17 #4700 @dshean
Hi @kthyng.  Any word from @jreadey?  I think he may be deep into the dev side, and may not be able to prioritize reviewing a submission like this.
Maybe time to ping others on the list?  I'm still hopeful that we can turn around revisions before end of 2022.  Thanks!
