### 2022-11-17 #4700 @kthyng
Hi @betolink! Would you be willing to review this submission for JOSS? The submission under consideration is SlideRule, a toolset for researchers leveraging the NASA ICESat-2 mission.
