### 2022-11-30 #4982 @editorialbot
[REVIEW]: SlideRule: Enabling rapid, scalable, open science for the NASA ICESat-2 mission and beyond
**Submitting author:** @dshean (<a href="http://orcid.org/0000-0003-3840-3860">David Shean</a>)
**Repository:** https://github.com/ICESat2-SlideRule/paper
**Branch with paper.md** (empty if default branch): main
**Version:** v1.5.8
