### 2022-12-29 ICESat2-SlideRule/paper#2 @jhkennedy
@jhkennedy's review for JOSS
I am collecting my thoughts/suggestions for openjournals/joss-reviews#498 here. Since this is a review of a collection of software and services, I may open more targeted discussions/issues/PR in the associated repos.

Note: This might be more appropriate as a discussion, but this repository does not have discussions enabled. Feel free to move it to a discussion if you like.

---

## Summary

Overall, I am really impressed by Sliderule and all the work that has gone into it. It clearly fulfills a very valuable need in the community and does so exceedingly well. Coming in from the outside, I found the services, documentation, and code very approachable and could get up and running pretty quickly with only a few minor hitches. There are some areas of the manuscript itself that could be improved, however, so **I am going to recommend accepting this manuscript with minor revisions.**

## Detailed Review

Here I'll go through [my review checklist](https://github.com/openjournals/joss-reviews/issues/4982#issuecomment-1332749544) and add discussion for any items I haven't already checked off -- if it's checked, I didn't feel the need to (further) discuss anything.

### General: Contribution and authorship

Overall, the authorship looks appropriate, but I am curious why Eric Lidwa is not included in the author list? It appears he had made substantial contributions even before this review was started in August (!) and has since become quite prolific on the project (e.g., [#2 contributor for sliderule](https://github.com/ICESat2-SlideRule/sliderule/graphs/contributors) and [#4 contributor for sliderule-python](https://github.com/ICESat2-SlideRule/sliderule-python/graphs/contributors))

### Documentation: Installation Instructions

#### sliderule

Following the README it was pretty straightforward to get to a successful build and install, but from there I never really got anything working. `sliderule scripts/selftests/test_runner.lua` would work and pass, but `sliderule scripts/apps/server.lua <config.json>` appears to need a `config.json` and I have *no idea* what that should be or look like.

Since there are docker instructions right in this repo (but no Dockerfile?) it'd be nice to provide a container for released versions a user/developer could pull down and run.

Other very minor thoughts:
* `make help` is super great! You should mention it in the README
* The Docker sections could probably just link out to the docker install docs instead of maintaining a copy yourself
* Most everything worked just the same for Ubuntu 22.04
* In the prerequisites section, I really had no idea what packages were available and why/when I would want them. I basically ignored the package dependency step entirely. It might be worth having an overview of the packages in a `packages/README.md`

### Documentation: Example usage

#### Viola Demo

This is fantastic overall. My only minor quibble is the controls require scrolling away from the map, but there's a lot of blank space to the right of the controls. If possible, it'd be nice to have those as a row along the plot instead of stacked in a column.

#### Sliderule-docs

In the [Getting started](https://slideruleearth.io/rtd/getting_started/Getting-Started.html) section of the user guide, I'd prefer "Examples" to be before "Project Map" so that when I click the "next" button at the bottom of the page, the examples appear after the basic usage.

In the [Making Your First Request](https://slideruleearth.io/rtd/tutorials/user/first_request.html) *user* tutorial:
  - I was [(also) confused](https://github.com/ICESat2-SlideRule/sliderule-python/issues/118#issuecomment-1340110311) by the netrc/auth warnings throw; maybe leaving `verbose=False` for this one would be better. I also think the error messages could be improved:
    - for `WARNING:sliderule.sliderule:Failed to retrieve username and password from netrc file: 'ps.slideruleearth.io'`, I expect a file path to come after the colon as written -- adding `... from netrc file for machine:` would be more clear
    - for `ERROR:sliderule.sliderule:Unable to authenticate user jhkennedy to https://ps.slideruleearth.io/api/org_token/`, I went down the path of hitting <https://ps.slideruleearth.io/api/org_token/>, determining the public org is `sliderule`, and finally getting the 403 Forbidden `"detail": "jhkennedy is NOT a member of sliderule"` message. Unfortunately, this "detail" is not passed to the user as part of the warning.
  - I am surprised this basic tutorial expects a development install? The other tutorials just expect you to have the sliderule client installed; I would expect the same here.

#### Sliderule-python

+1 to: https://github.com/ICESat2-SlideRule/sliderule-python/issues/119

Also, `verbose=True` is in the README here as well, but I'd prefer it not to be (see above).

### Documentation: Community Guidelines

It'd be nice to include your [Contributing Guidelines](https://slideruleearth.io/rtd/user_guide/Contributing.html) in your [.github repo](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file) so that they are easily found and reported by GitHub. e.g.: https://github.com/ICESat2-SlideRule/sliderule-python/community

Outside contributions to `sliderule-python` are likely hampered by: ICESat2-SlideRule/sliderule-python#121

### Software paper: Quality of writing

Overall very good! I have some minor suggestions that I think are best left here in this format.

I would like to see the version numbers stated for `sliderule`, `sliderule-python` and `sliderule-docs` -- this will help readers in the future understand what version was reviewed and help me as a reviewer separate out current development work from the latest release. Notably, since version numbers for all three seemed to be synced, I've been mostly looking at `v1.5.8` and subsequent development.

I found the transition to the ["ICESat-2 packages"](https://github.com/ICESat2-SlideRule/paper/blob/main/paper/paper.md#icesat-2-packages) a bit abrupt and without the necessary context.

It might be worth noting this API is currently a synchronous service (the user connection stays open and results stream in) as opposed to an async service (the user submits a request and comes back later for data) such as ASF HyP3 -- synchronous *is* a better user experience as long as the time to results is "short". Likewise, are there associated performance goals with the service? And are async data requests expected in the future or excluded from this service?

---

### Other things

I noticed sliderule is undergoing a transition from <http://icesat2sliderule.org/> to <https://slideruleearth.io>. There are a few broken links around because of it.
* ICESat2-SlideRule/sliderule#155
* ICESat2-SlideRule/paper#1

It might be worth looking through all the hits for the old `.org` in the GitHub org:
https://github.com/search?q=org%3AICESat2-SlideRule%20sliderule.org&type=code
