### 2022-12-29 #4982 @jhkennedy
@kthyng I have finished my (initial?) review of sliderule, and provided my detailed review in ICESat2-SlideRule/paper#2.

Overall, I am really impressed by Sliderule and all the work that has gone into it. It clearly fulfills a very valuable need in the community and does so exceedingly well. Coming in from the outside, I found the services, documentation, and code very approachable and could get up and running pretty quickly with only a few minor hitches. There are some areas of the manuscript itself that could be improved, however, so **I  recommend accepting this manuscript with minor revisions.**

I am mostly concerned with the things in the subsection "Software paper: Quality of writing" of that issue (very bottom) which should be quick to address and would be happy to recommend accept when addressed.
