### 2022-12-29 ICESat2-SlideRule/paper#2 @jhkennedy
And tagging @dshean @jpswinski to ensure they see this.
