### 2023-01-04 #4982 @dshean
Thanks @jhkennedy and @betolink for your detailed reviews!

We have already started preparing responses and will follow up in the discussion thread (https://github.com/ICESat2-SlideRule/sliderule/discussions/154) and the JOSS review comment thread here.

@kthyng can you provide additional guidance on status of this review and required next steps in the JOSS review process, preferred formats, etc?  A bit different than traditional academic journals, but I assume a detailed response to each comment is required here?  Or are separate responses in PRs and associated github discussion sufficient?
