### 2023-01-10 ICESat2-SlideRule/paper#2 @jpswinski
Thank you for the thorough review of our repos!  I appreciate that you took the time to dig as deep as you did and provide the feedback that you did.

### Documentation: Installation Instructions

I made the recommended changes to the README documentation in the sliderule repo at cd7780b, and 04760a8.  There are also some changes related to handling endianness.  When updating the readme per the feedback you provided, I realized that the configuration support for selecting endianness was not great, so I cleaned that up.
* I added a note on running `make help` to the README
* I removed the instructions for installing docker from the README.
* Ubuntu 22.04 is a bit of a conundrum for us at the moment and we are currently not supporting it.  Everything will build and run fine, but the performance takes a 50% to 100% hit when running large processing runs.  We lost a week of time trying to figure it out (mostly because I incorrectly thought it had to do with an AWS SDK update), but have not had the time to revisit and get to the bottom of it.
* I added a packages/README.md and plugins/README.md

### Documentation: Example usage

* Voila page layout - we fully agree.  The initial scope of the demo was to have a quick way to run ipywidgets without needing to run a jupyter notebook.  We considered several options for interaction, including Panel, Cesium, and Voila. As the scope has grown, we’ve concluded that voila is not the right solution, so we are in the process of designing (and then implementing) a more sophisticated web-based user interface  that will really be a software tool in its own right.
* Getting Started - as suggested, I moved examples up to be right after the getting started page.
* Making Your First Request - I changed the netrc warning message as suggested, and now also only issue the warning when attempting to access a private cluster (so for the public sliderule cluster, the warning will not be issued).  I changed the tutorial to specify verbose=False as suggested.
  * Regarding authentication - we are currently in the process of migrating our entire system to a new architecture.  Our old architecture had no authentication anywhere.  In our new architecture, we have a concept of private clusters and public clusters.  We have a public cluster running all the time at https://sliderule.slideruleearth.io/, and by default everything uses this public cluster.  When using a public cluster, there is no need to authenticate (though as developers we are the only ones that are members of the public cluster which does give us access to some control APIs when we do authenticate).  For private clusters, authentication allows access to the cluster.  They are reached at https://{org}.slideruleearth.io, where {org} is the name of the organization associated with the cluster.  Very little of this is reflected in our documentation, and we are only just now beginning to support our first groups interested in private clusters.
* Developer Install Required - I was thinking using `conda env create -f environment.yml` was the safest option  for new users because it creates a new environment and installs everything needed for any examples after the tutorial.
README.md - I changed verbose to False as suggested

### Documentation: Community Guidelines

* Contributor Guidelines - I created issue ICESat2-Sliderule/sliderule#170 to add a CONTRIBUTING file to our repos.  One question I have is if you have any recommendations on how to handle the fact that we have multiple repos.  Do you have a copy of the file in each repo?  Do you point to the website in each file, or have the full text?  I’ve found managing the multiple SlideRule repos on GitHub to be challenging (they definitely started out as a Repo-First system and are only slowly moving towards an Org-First approach).  I’d be glad for any lessons learned you have on this.
* [ICESat2-SlideRule/sliderule-python#121](https://github.com/ICESat2-SlideRule/sliderule-python/issues/121) - see response in comments of issue
