### 2023-01-10 ICESat2-SlideRule/paper#2 @dshean
Thanks again @jhkennedy.  Here are additional responses regarding the paper:

> Overall, the authorship looks appropriate, but I am curious why Eric Lidwa is not included in the author list? It appears he had made substantial contributions even before this review was started in August (!) and has since become quite prolific on the project (e.g.,[ ICESat2-SlideRule/sliderule-python#2 contributor for sliderule](https://github.com/ICESat2-SlideRule/sliderule/graphs/contributors) and[ ICESat2-SlideRule/sliderule-python#4 contributor for sliderule-python](https://github.com/ICESat2-SlideRule/sliderule-python/graphs/contributors))

This paper was initially prepared to summarize the initial ~1.5 years of the SlideRule project, before Eric joined the team for the ongoing ICESat-2 mission extension project.  But considering the time that has passed since initial submission and Eric’s subsequent contributions, we added him to the coauthor list. It is worth noting that coauthor @cugarteblair has also made significant contributions to the SlideRule provisioning system development (https://github.com/ICESat2-SlideRule/sliderule-ps-server/graphs/contributors and https://github.com/ICESat2-SlideRule/sliderule-ps-web/graphs/contributors), which was not covered in this paper.
We had several internal discussions about the effort required vs. impact of an academic paper, most valuable content to include, and author order.  We have all made significant contributions to the project over the years and it is challenging to rank or capture these contributions in a snapshot author order for an atypical journal like JOSS, even with systems like CRediT. The reality is that academic publications are not a priority for the SlideRule development team, but they are essential for the SlideRule science/management team, and the latter led the JOSS paper effort.

>Software paper: Quality of writing: I would like to see the version numbers stated for sliderule, sliderule-python and sliderule-docs -- this will help readers in the future understand what version was reviewed and help me as a reviewer separate out current development work from the latest release. Notably, since version numbers for all three seemed to be synced, I've been mostly looking at v1.5.8 and subsequent development.

Good suggestion to document version number for paper and reviews.  We added version number to the Sliderule project section:
“As of version 1.5.8, the SlideRule project includes the following repositories:”

>I found the transition to the[ "ICESat-2 packages"](https://github.com/ICESat2-SlideRule/paper/blob/main/paper/paper.md#icesat-2-packages) a bit abrupt and without the necessary context.

Agreed. This was included as part of the “state of the field” section.  We modified the section head to “Currently available ICESat-2 processing packages” and added some introductory sentences “Several existing projects offer software and/or APIs to process, analyze and visualize ICESat-2 data products. We briefly summarize these efforts to provide context and justification for the ICESat-2 SlideRule project.”

>It might be worth noting this API is currently a synchronous service (the user connection stays open and results stream in) as opposed to an async service (the user submits a request and comes back later for data) such as ASF HyP3 -- synchronous is a better user experience as long as the time to results is "short".

Good point! We added the following text to the “Sliderule server framework” section:

“The SlideRule service was designed for synchronous processing - the client connection remains open after a request is submitted, and results are streamed back to the engaged user in near-real-time. This model is preferable over asynchronous processing, where requests are queued and users are notified to retrieve results at a later time.”

> Likewise, are there associated performance goals with the service? And are async data requests expected in the future or excluded from this service?

We don’t currently have specific performance goals, and the actual time delay depends on the size of the request, number of ATL03 granules and the configuration of the processing cluster. In general, we aim to deliver results as fast as possible, so the user stays engaged and any delay does not interrupt the interactive testing and development process.
