### 2023-01-10 #4982 @dshean
OK @kthyng, @jhkennedy and @betolink.  I believe we addressed all reviewer comments in respective threads:

- https://github.com/ICESat2-SlideRule/sliderule/discussions/154#discussioncomment-4647824
- https://github.com/ICESat2-SlideRule/paper/issues/2#issuecomment-1377602593
- https://github.com/ICESat2-SlideRule/paper/issues/2#issuecomment-1377606461

Changes to the revised paper are here: https://github.com/ICESat2-SlideRule/paper/commit/19ac7d0dd1cc467a8df659f215464c8bf7f459eb.  @kthyng, should we merge this into main and rebuild pdf?

Thanks again for your thorough review, and please let us know if there are any lingering items that we need to address before acceptance!
