### 2023-01-11 #4982 @betolink
Hi @kthyng, I'm happy with the response I got from the SlideRule team to my [questions](https://github.com/ICESat2-SlideRule/sliderule/discussions/154), I anticipate there will be some follow up but overall I'm impressed by the quality of their tools and the documentation. I recommend accepting this manuscript and look forward to future work on this much needed framework for science in the cloud.
