### 2023-01-12 ICESat2-SlideRule/paper#2 @jhkennedy
Overal, I'm happy with this feedback! I'm going to close this issue and continue some of the related discussions in the appropriate issues ( https://github.com/ICESat2-SlideRule/sliderule-python/issues/121, https://github.com/ICESat2-SlideRule/sliderule/issues/170 )

---

> Ubuntu 22.04 is a bit of a conundrum for us at the moment and we are currently not supporting it. Everything will build and run fine, but the performance takes a 50% to 100% hit when running large processing runs. We lost a week of time trying to figure it out (mostly because I incorrectly thought it had to do with an AWS SDK update), but have not had the time to revisit and get to the bottom of it.

I feel that! We've had lots of similar "fun" when upgrading OSs and critical software dependencies.
