### 2023-01-12 #4982 @jhkennedy
@kthyng and @dshean, I'm happy with the responses I got from the SlideRule team to my [review](https://github.com/ICESat2-SlideRule/paper/issues/2) and have closed the issue. While I expect we will continue a couple of conversations,  **I recommend accepting this manuscript** and moving forward with publication as I don't think those conversations affect this manuscript.

I really enjoyed the review process (hopefully, I wasn't too thorough), and look forward to all the amazing work planned by the SlideRule team.
