### 2023-01-12 #4982 @kthyng
Ok then we have both reviewers confirmed and all boxes checked! And yes discussion can certainly continue on.

@dshean I'll let you merge and regenerate the pdf — please make sure that the paper is exactly how you it should look (especially the references) since after I read through it, if I have no comments then I will accept it.

Other steps for this point for @dshean:
* After the code is merged, tag a new release for the present state and report the version back here
* Upload an archive of your code to a place like Zenodo and report back the DOI here. Please alter the metadata of the archive so that the title and author list exactly match your JOSS submission.
