+++ paper/paper.md
@@ -53,12 +53,12 @@ ICESat-2 launched in September 2018 as the follow-on mission for the original IC
 
 ICESat-2 carries the Advanced Topographic Laser Altimeter System (ATLAS), which records the two-way travel time of transmitted photons from six beams [@neumann2019ice]. The Level 2A Geolocated Photon Data Product (ATL03) provides unique latitude, longitude, height and timing of each photon, along with many other attributes, including basic classification as a return from the Earth's surface vs. background photons from sunlight or instrument noise.
 
-The ATL03 data granules are stored as ~1-2 GB HDF5-format files containing ~20-100 million records. This data volume and complexity impedes new users and slows the speed of algorithm development. The ICESat-2 project generates many [higher-level products](https://nsidc.org/data/icesat-2/products) that reduce the ATL03 data using established processing algorithms. For example, the ATL06 product [@smith2019land] provides high-precision estimates of surface height at 40-meter resolution, using parameters appropriate for flat, highly reflective polar snow surfaces. Similar to ATL03, the ATL06 products contain a large number of parameters describing each height measurement that may not be relevant to all users. Furthermore, standard ATL06 products are only produced over glaciers and ice sheets, and the default algorithm parameters (such as 40-meter spacing) are not optimal for complex land surfaces or vegetation. 
+The ATL03 data granules are stored as ~1-2 GB HDF5-format files containing ~20-100 million records. This data volume and complexity impedes new users and slows the speed of algorithm development. The ICESat-2 project team generates many [higher-level products](https://nsidc.org/data/icesat-2/products) that reduce the ATL03 data using established processing algorithms. For example, the ATL06 product [@smith2019land] provides high-precision estimates of surface height at 40-meter resolution, using parameters appropriate for flat, highly reflective polar snow surfaces. Similar to ATL03, the ATL06 products contain a large number of parameters describing each height measurement that may not be relevant to all users. Furthermore, standard ATL06 products are only produced over glaciers and ice sheets, and the default algorithm parameters (such as 20-meter posting) are not optimal for complex land surfaces or vegetation. 
 
 `SlideRule` offers a solution to these issues, allowing users to create on-demand products with the vetted ATL06 algorithm, but with adjustable parameters and photon classification strategies tailored to the characteristics of their application and study sites.
 
 ## State of the field
-The current paradigm for ICESat-2 data access involves downloading large volumes of standard data products from a NASA Distributed Active Archive Center (DAAC), then writing custom routines to prepare those products for analysis. The National Snow and Ice Data Center (NSIDC) offers data discovery and limited subsetting services, allowing users to request and download products for a user-specified geographic area with a user-defined subset of returned variables [@atl03_nsidc]. Even with these subsetting services, the full workflow to request, stage, and download hundreds of products can take several minutes to hours, especially for larger areas, and these services do not currently support custom server-side data processing.
+The current paradigm for ICESat-2 data access involves downloading large volumes of standard data products from a NASA Distributed Active Archive Center (DAAC), then writing custom routines to prepare those products for analysis. The National Snow and Ice Data Center (NSIDC) offers data discovery and limited subsetting services, allowing users to request and download products for a user-specified geographic area with a user-defined subset of returned variables. Even with these subsetting services, the full workflow to request, stage, and download hundreds of products can take several minutes to hours, especially for larger areas, and these services do not currently support custom server-side data processing.
 
 ### On-demand science data processing
[…]
