+++ paper/paper.bib
@@ -122,7 +122,7 @@ author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalat
 author =    {Scheick, Jessica and others},
 organization = {icesat2py},
 title =     {{icepyx: Python} tools for obtaining and working with {ICESat-2} data},
-year =      {2019--},
+year =      {2019},
 url = "https://github.com/icesat2py/icepyx"
 }
 
