### 2023-01-12 #4982 @dshean
Thanks @kthyng!  I merged the paper revision PR and made a few minor corrections after another final review.  I think we are good to go with latest main branch in the paper repo: https://github.com/ICESat2-SlideRule/paper.
