### 2023-01-12 #4982 @dshean
The reviews and paper were tagged at v1.5.8 (thanks to @jhkennedy's suggestion).  Here are relevant release tags and DOIs from Zenodo for the code and docs:

https://github.com/ICESat2-SlideRule/sliderule-python/tree/v1.5.8 (https://doi.org/10.5281/zenodo.7477718)
https://github.com/ICESat2-SlideRule/sliderule/tree/v1.5.8 (https://doi.org/10.5281/zenodo.7477717)
https://github.com/ICESat2-SlideRule/sliderule-docs/tree/v1.5.8 (no Zenodo record for docs)
