### 2023-01-16 #4982 @kthyng
@dshean Thanks. For this sort of review with multiple associated repositories, could you make a new, separate Zenodo archive that contains a snapshot of all the related repos? The paper repo can be included but isn't so important for archiving (since we can see the paper) but the other 3 should be included. For this combined repo please adjust the metadata.
