### 2023-01-16 #4982 @dshean
Hi @kthyng.

As requested, I created a single Zenodo record containing the snapshot of the 3 repositories and the paper: https://doi.org/10.5281/zenodo.7542007.  Please let me know if I should modify the description or other metadata.

I will update with final details for the paper after acceptance and publication.  Thank you.
