### 2023-01-17 #4982 @kthyng
Comments on paper:
* typo last paragraph in Summary "developent"
* Why are there question marks after the reviewer github handles in acknowledgments?
