### 2023-01-17 79228aa
Editor review: Fix typo and github handles
